Change Logic
Resolution of bug payments, the payment exceeds has a bad validation.

+++ billing-service/src/main/java/com/billing/service/billing_service/repositorys/BillsRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.billing.service.billing_service.domain.Bills;
 
-public interface BillsRepository extends JpaRepository<Bills, Integer>{
+public interface BillsRepository extends JpaRepository<Bills, Long>{
 }
